dapp/encryption/requirements.txt
Flask
pycryptodome
flask-cors